:truck: Move filter classes to new folder
